fix port issues, prepare for building
added:
 obs_build/
 obs_build/debian.changelog
 obs_build/debian.control
 obs_build/debian.rules
 obs_build/g2cd-Debian_5.0.dsc
 obs_build/g2cd-Debian_Etch.dsc
 obs_build/g2cd-xUbuntu_10.04.dsc
 obs_build/g2cd-xUbuntu_8.04.dsc
 obs_build/g2cd-xUbuntu_9.04.dsc
 obs_build/g2cd-xUbuntu_9.10.dsc
 obs_build/g2cd.spec
	Online Build Service control files.
	Yes, now we have builds for a truckload of systems (which are
	not really usable) at:
	https://build.opensuse.org/package/show?package=g2cd&project=home%3Akaffeemonster%3Ag2cd
	or, who wants files:
	http://download.opensuse.org/repositories/home:/kaffeemonster:/g2cd/

 lib/my_bitops.c
 lib/recv_buff.c
 G2QueryKey.c
 G2ConRegistry.c
	include config to fix comile on systems with "funny" header

 G2HeaderFieldsSort.c
	fox 64 bit, the field width is only int

 configure.ac
 G2MainServer.c
	test for PR_SET_NAME, old systems do not have it

 bin2o.c
	add a .note.GNU-stack section so these obj do not set writeable
	stacks, so our complete binary gets NX.
	Lets see which system will break with that...

 lib/generic/little_endian.h
	bad typo, missing parenthesis, wrong func, only saw on old systems...

 Makefile
	add files
	create .gz by default, thanks Debian...
	link ccdrv first with ncurses, then curses, then fal back to helper
	change helper to include one pair more of quoting, even if it
	work here, i'm sure it is broken, because other shells do not eat
	it, i think to remember.

 .bzrignore
	add *.gz to the ignore list, wonder i didn't include that already

+++ G2HeaderFieldsSort.c
@@ -188,8 +188,8 @@ int main(int argc, char *argv[])
 		for(i = 0; i < anum(hdr_fields); i++)
 		{
 			fprintf(out, "\t{%*s, str_size(%*s), %*s }, /* %s */\n",
-			       max_f_name, hdr_fields[i]->f_name, max_ktxt, hdr_fields[i]->ktxt,
-			       max_ktxt, hdr_fields[i]->ktxt, hdr_fields[i]->txt);
+			       (int)max_f_name, hdr_fields[i]->f_name, (int)max_ktxt, hdr_fields[i]->ktxt,
+			       (int)max_ktxt, hdr_fields[i]->ktxt, hdr_fields[i]->txt);
 		}
 
 		fputs("};\n\n/* EOF */\n", out);
